+++ x-pack/plugins/index_lifecycle_management/__jest__/client_integration/edit_policy/edit_policy.helpers.tsx
@@ -183,13 +183,6 @@ export const setup = async (arg?: {
 
   const enable = (phase: Phases) => createFormToggleAction(`enablePhaseSwitch-${phase}`);
 
-  const showDataAllocationOptions = (phase: Phases) => () => {
-    act(() => {
-      find(`${phase}-dataTierAllocationControls.dataTierSelect`).simulate('click');
-    });
-    component.update();
-  };
-
   const createMinAgeActions = (phase: Phases) => {
     return {
       hasMinAgeInput: () => exists(`${phase}-selectedMinimumAge`),
@@ -384,7 +377,6 @@ export const setup = async (arg?: {
       },
       warm: {
         enable: enable('warm'),
-        showDataAllocationOptions: showDataAllocationOptions('warm'),
         ...createMinAgeActions('warm'),
         setReplicas: setReplicas('warm'),
         hasErrorIndicator: () => exists('phaseErrorIndicator-warm'),
@@ -396,7 +388,6 @@ export const setup = async (arg?: {
       },
       cold: {
         enable: enable('cold'),
-        showDataAllocationOptions: showDataAllocationOptions('cold'),
         ...createMinAgeActions('cold'),
         setReplicas: setReplicas('cold'),
         setFreeze,
