+++ x-pack/plugins/index_lifecycle_management/__jest__/client_integration/edit_policy/serialization/policy_serialization.test.ts
@@ -1,0 +1,426 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { act } from 'react-dom/test-utils';
+import { setupEnvironment } from '../../helpers/setup_environment';
+import {
+  getDefaultHotPhasePolicy,
+  POLICY_WITH_INCLUDE_EXCLUDE,
+  POLICY_WITH_KNOWN_AND_UNKNOWN_FIELDS,
+} from '../constants';
+import { EditPolicyTestBed, setup } from '../edit_policy.helpers';
+
+describe('<EditPolicy /> serialization', () => {
+  let testBed: EditPolicyTestBed;
+  const { server, httpRequestsMockHelpers } = setupEnvironment();
+
+  afterAll(() => {
+    server.restore();
+  });
+
+  beforeEach(async () => {
+    httpRequestsMockHelpers.setLoadPolicies([getDefaultHotPhasePolicy('my_policy')]);
+    httpRequestsMockHelpers.setLoadSnapshotPolicies([]);
+    httpRequestsMockHelpers.setListNodes({
+      nodesByRoles: {},
+      nodesByAttributes: { test: ['123'] },
+      isUsingDeprecatedDataRoleConfig: false,
+    });
+
+    await act(async () => {
+      testBed = await setup();
+    });
+
+    const { component } = testBed;
+    component.update();
+  });
+
+  describe('top level form', () => {
+    /**
+     * We assume that policies that populate this form are loaded directly from ES and so
+     * are valid according to ES. There may be settings in the policy created through the ILM
+     * API that the UI does not cater for, like the unfollow action. We do not want to overwrite
+     * the configuration for these actions in the UI.
+     */
+    it('preserves policy settings it did not configure', async () => {
+      httpRequestsMockHelpers.setLoadPolicies([POLICY_WITH_KNOWN_AND_UNKNOWN_FIELDS]);
+      await act(async () => {
+        testBed = await setup();
+      });
+
+      const { component, actions } = testBed;
+      component.update();
+
+      // Set max docs to test whether we keep the unknown fields in that object after serializing
+      await actions.hot.setMaxDocs('1000');
+      // Remove the delete phase to ensure that we also correctly remove data
+      await actions.delete.disablePhase();
+      await actions.savePolicy();
+
+      const latestRequest = server.requests[server.requests.length - 1];
+      const entirePolicy = JSON.parse(JSON.parse(latestRequest.requestBody).body);
+
+      expect(entirePolicy).toEqual({
+        foo: 'bar', // Made up value
+        name: 'my_policy',
+        phases: {
+          hot: {
+            actions: {
+              rollover: {
+                max_docs: 1000,
+                max_size: '50gb',
+                unknown_setting: 123, // Made up setting that should stay preserved
+              },
+            },
+            min_age: '0ms',
+          },
+          warm: {
+            actions: {
+              my_unfollow_action: {}, // Made up action
+              set_priority: {
+                priority: 22,
+                unknown_setting: true,
+              },
+            },
+            min_age: '0d',
+          },
+        },
+      });
+    });
+  });
+
+  describe('hot phase', () => {
+    test('setting all values', async () => {
+      const { actions } = testBed;
+
+      await actions.hot.toggleDefaultRollover(false);
+      await actions.hot.setMaxSize('123', 'mb');
+      await actions.hot.setMaxDocs('123');
+      await actions.hot.setMaxAge('123', 'h');
+      await actions.hot.toggleForceMerge(true);
+      await actions.hot.setForcemergeSegmentsCount('123');
+      await actions.hot.setBestCompression(true);
+      await actions.hot.toggleShrink(true);
+      await actions.hot.setShrink('2');
+      await actions.hot.toggleReadonly(true);
+      await actions.hot.toggleIndexPriority(true);
+      await actions.hot.setIndexPriority('123');
+
+      await actions.savePolicy();
+      const latestRequest = server.requests[server.requests.length - 1];
+      const entirePolicy = JSON.parse(JSON.parse(latestRequest.requestBody).body);
+      expect(entirePolicy).toMatchInlineSnapshot(`
+          Object {
+            "name": "my_policy",
+            "phases": Object {
+              "hot": Object {
+                "actions": Object {
+                  "forcemerge": Object {
+                    "index_codec": "best_compression",
+                    "max_num_segments": 123,
+                  },
+                  "readonly": Object {},
+                  "rollover": Object {
+                    "max_age": "123h",
+                    "max_docs": 123,
+                    "max_size": "123mb",
+                  },
+                  "set_priority": Object {
+                    "priority": 123,
+                  },
+                  "shrink": Object {
+                    "number_of_shards": 2,
+                  },
+                },
+                "min_age": "0ms",
+              },
+            },
+          }
+        `);
+    });
+
+    test('setting searchable snapshot', async () => {
+      const { actions } = testBed;
+
+      await actions.hot.setSearchableSnapshot('my-repo');
+
+      await actions.savePolicy();
+      const latestRequest = server.requests[server.requests.length - 1];
+      const entirePolicy = JSON.parse(JSON.parse(latestRequest.requestBody).body);
+      expect(entirePolicy.phases.hot.actions.searchable_snapshot.snapshot_repository).toBe(
+        'my-repo'
+      );
+    });
+
+    test('disabling rollover', async () => {
+      const { actions } = testBed;
+      await actions.hot.toggleDefaultRollover(false);
+      await actions.hot.toggleRollover(false);
+      await actions.savePolicy();
+      const latestRequest = server.requests[server.requests.length - 1];
+      const policy = JSON.parse(JSON.parse(latestRequest.requestBody).body);
+      const hotActions = policy.phases.hot.actions;
+      const rolloverAction = hotActions.rollover;
+      expect(rolloverAction).toBe(undefined);
+      expect(hotActions).toMatchInlineSnapshot(`Object {}`);
+    });
+  });
+
+  describe('warm phase', () => {
+    beforeEach(async () => {
+      httpRequestsMockHelpers.setLoadPolicies([getDefaultHotPhasePolicy('my_policy')]);
+      httpRequestsMockHelpers.setListNodes({
+        nodesByRoles: {},
+        nodesByAttributes: { test: ['123'] },
+        isUsingDeprecatedDataRoleConfig: false,
+      });
+      httpRequestsMockHelpers.setLoadSnapshotPolicies([]);
+
+      await act(async () => {
+        testBed = await setup();
+      });
+
+      const { component } = testBed;
+      component.update();
+    });
+
+    test('default values', async () => {
+      const { actions } = testBed;
+      await actions.warm.enable(true);
+      await actions.savePolicy();
+      const latestRequest = server.requests[server.requests.length - 1];
+      const warmPhase = JSON.parse(JSON.parse(latestRequest.requestBody).body).phases.warm;
+      expect(warmPhase).toMatchInlineSnapshot(`
+          Object {
+            "actions": Object {
+              "set_priority": Object {
+                "priority": 50,
+              },
+            },
+            "min_age": "0d",
+          }
+        `);
+    });
+
+    test('setting all values', async () => {
+      const { actions } = testBed;
+      await actions.warm.enable(true);
+      await actions.warm.setDataAllocation('node_attrs');
+      await actions.warm.setSelectedNodeAttribute('test:123');
+      await actions.warm.setReplicas('123');
+      await actions.warm.toggleShrink(true);
+      await actions.warm.setShrink('123');
+      await actions.warm.toggleForceMerge(true);
+      await actions.warm.setForcemergeSegmentsCount('123');
+      await actions.warm.setBestCompression(true);
+      await actions.warm.toggleReadonly(true);
+      await actions.warm.setIndexPriority('123');
+      await actions.savePolicy();
+      const latestRequest = server.requests[server.requests.length - 1];
+      const entirePolicy = JSON.parse(JSON.parse(latestRequest.requestBody).body);
+      // Check shape of entire policy
+      expect(entirePolicy).toMatchInlineSnapshot(`
+          Object {
+            "name": "my_policy",
+            "phases": Object {
+              "hot": Object {
+                "actions": Object {
+                  "rollover": Object {
+                    "max_age": "30d",
+                    "max_size": "50gb",
+                  },
+                },
+                "min_age": "0ms",
+              },
+              "warm": Object {
+                "actions": Object {
+                  "allocate": Object {
+                    "number_of_replicas": 123,
+                    "require": Object {
+                      "test": "123",
+                    },
+                  },
+                  "forcemerge": Object {
+                    "index_codec": "best_compression",
+                    "max_num_segments": 123,
+                  },
+                  "readonly": Object {},
+                  "set_priority": Object {
+                    "priority": 123,
+                  },
+                  "shrink": Object {
+                    "number_of_shards": 123,
+                  },
+                },
+                "min_age": "0d",
+              },
+            },
+          }
+        `);
+    });
+
+    describe('policy with include and exclude', () => {
+      beforeEach(async () => {
+        httpRequestsMockHelpers.setLoadPolicies([POLICY_WITH_INCLUDE_EXCLUDE]);
+        httpRequestsMockHelpers.setListNodes({
+          nodesByRoles: {},
+          nodesByAttributes: { test: ['123'] },
+          isUsingDeprecatedDataRoleConfig: false,
+        });
+        httpRequestsMockHelpers.setLoadSnapshotPolicies([]);
+
+        await act(async () => {
+          testBed = await setup();
+        });
+
+        const { component } = testBed;
+        component.update();
+      });
+
+      test('preserves include, exclude allocation settings', async () => {
+        const { actions } = testBed;
+        await actions.warm.setDataAllocation('node_attrs');
+        await actions.warm.setSelectedNodeAttribute('test:123');
+        await actions.savePolicy();
+        const latestRequest = server.requests[server.requests.length - 1];
+        const warmPhaseAllocate = JSON.parse(JSON.parse(latestRequest.requestBody).body).phases.warm
+          .actions.allocate;
+        expect(warmPhaseAllocate).toMatchInlineSnapshot(`
+          Object {
+            "exclude": Object {
+              "def": "456",
+            },
+            "include": Object {
+              "abc": "123",
+            },
+            "require": Object {
+              "test": "123",
+            },
+          }
+        `);
+      });
+    });
+  });
+
+  describe('cold phase', () => {
+    beforeEach(async () => {
+      httpRequestsMockHelpers.setLoadPolicies([getDefaultHotPhasePolicy('my_policy')]);
+      httpRequestsMockHelpers.setListNodes({
+        nodesByRoles: {},
+        nodesByAttributes: { test: ['123'] },
+        isUsingDeprecatedDataRoleConfig: false,
+      });
+      httpRequestsMockHelpers.setLoadSnapshotPolicies([]);
+
+      await act(async () => {
+        testBed = await setup();
+      });
+
+      const { component } = testBed;
+      component.update();
+    });
+
+    test('default values', async () => {
+      const { actions } = testBed;
+
+      await actions.cold.enable(true);
+      await actions.savePolicy();
+      const latestRequest = server.requests[server.requests.length - 1];
+      const entirePolicy = JSON.parse(JSON.parse(latestRequest.requestBody).body);
+      expect(entirePolicy.phases.cold).toMatchInlineSnapshot(`
+          Object {
+            "actions": Object {
+              "set_priority": Object {
+                "priority": 0,
+              },
+            },
+            "min_age": "0d",
+          }
+        `);
+    });
+
+    test('setting all values, excluding searchable snapshot', async () => {
+      const { actions } = testBed;
+
+      await actions.cold.enable(true);
+      await actions.cold.setMinAgeValue('123');
+      await actions.cold.setMinAgeUnits('s');
+      await actions.cold.setDataAllocation('node_attrs');
+      await actions.cold.setSelectedNodeAttribute('test:123');
+      await actions.cold.setReplicas('123');
+      await actions.cold.setFreeze(true);
+      await actions.cold.setIndexPriority('123');
+
+      await actions.savePolicy();
+      const latestRequest = server.requests[server.requests.length - 1];
+      const entirePolicy = JSON.parse(JSON.parse(latestRequest.requestBody).body);
+
+      expect(entirePolicy).toMatchInlineSnapshot(`
+          Object {
+            "name": "my_policy",
+            "phases": Object {
+              "cold": Object {
+                "actions": Object {
+                  "allocate": Object {
+                    "number_of_replicas": 123,
+                    "require": Object {
+                      "test": "123",
+                    },
+                  },
+                  "freeze": Object {},
+                  "set_priority": Object {
+                    "priority": 123,
+                  },
+                },
+                "min_age": "123s",
+              },
+              "hot": Object {
+                "actions": Object {
+                  "rollover": Object {
+                    "max_age": "30d",
+                    "max_size": "50gb",
+                  },
+                },
+                "min_age": "0ms",
+              },
+            },
+          }
+        `);
+    });
+
+    // Setting searchable snapshot field disables setting replicas so we test this separately
+    test('setting searchable snapshot', async () => {
+      const { actions } = testBed;
+      await actions.cold.enable(true);
+      await actions.cold.setSearchableSnapshot('my-repo');
+      await actions.savePolicy();
+      const latestRequest2 = server.requests[server.requests.length - 1];
+      const entirePolicy2 = JSON.parse(JSON.parse(latestRequest2.requestBody).body);
+      expect(entirePolicy2.phases.cold.actions.searchable_snapshot.snapshot_repository).toEqual(
+        'my-repo'
+      );
+    });
+  });
+
+  test('delete phase', async () => {
+    const { actions } = testBed;
+    await actions.delete.enablePhase();
+    await actions.setWaitForSnapshotPolicy('test');
+    await actions.savePolicy();
+    const latestRequest = server.requests[server.requests.length - 1];
+    const entirePolicy = JSON.parse(JSON.parse(latestRequest.requestBody).body);
+    expect(entirePolicy.phases.delete).toEqual({
+      min_age: '365d',
+      actions: {
+        delete: {},
+        wait_for_snapshot: {
+          policy: 'test',
+        },
+      },
+    });
+  });
+});
